+++ setup.py
@@ -1,0 +1,9 @@
+"""
+Python Flight Mechanics Engine (PyFME).
+Copyright (c) AeroPython Development Team.
+Distributed under the terms of the MIT License.
+
+Title
+-----
+
+"""
